refactor(pages/query-tracker): do not import actions from reducers [YTFRONT-3814]
