+++ packages/ui/src/ui/pages/query-tracker/QueryResults/QueryResultActions/index.tsx
@@ -5,7 +5,7 @@ import React, {useCallback} from 'react';
 import {useDispatch, useSelector} from 'react-redux';
 import {RootState} from '../../../../store/reducers';
 import transposeIcon from '../../../../assets/img/svg/transpose.svg';
-import {SET_QUERY_RESULTS_SETTINGS} from '../../module/query_result/actions';
+import {SET_QUERY_RESULTS_SETTINGS} from '../../module/query-tracker-contants';
 import {QueryItem} from '../../module/api';
 import {QueryResultReadyState, QueryResultsViewMode} from '../../module/query_result/types';
 import {TableColumnsSelector} from './TableColumnsSelector';
