+++ packages/ui/src/ui/pages/query-tracker/module/queries_list/actions.ts
@@ -3,18 +3,20 @@ import {ThunkAction} from 'redux-thunk';
 import {ActionD} from '../../../../types';
 import {wrapApiPromiseByToaster} from '../../../../utils/utils';
 import {RootState} from '../../../../store/reducers';
-import {QueriesHistoryCursorDirection, QueryItem, loadQueriesList} from '../api';
+import {QueryItem, loadQueriesList} from '../api';
 import {QueriesListState} from './reducer';
 import {getQueriesList, getQueriesListCursorParams, getQueriesListFilterParams} from './selectors';
 import {QueriesListCursor, QueriesListFilter} from './types';
-import {UPDATE_QUERIES_LIST} from '../query-tracker-contants';
-
-export const LOAD_QUERIES_LIST_REQUEST = 'query-tracker/LOAD_QUERIES_LIST_REQUEST';
-export const LOAD_QUERIES_LIST_SUCCESS = 'query-tracker/LOAD_QUERIES_LIST_SUCCESS';
-export const LOAD_QUERIES_LIST_ERROR = 'query-tracker/LOAD_QUERIES_LIST_ERROR';
-export const SET_QUERIES_LIST_FILTER = 'query-tracker/SET_QUERIES_LIST_FILTER';
-export const SET_QUERIES_LIST_MODE = 'query-tracker/SET_QUERIES_LIST_MODE';
-export const SET_QUERIES_LIST_CURSOR = 'query-tracker/SET_QUERIES_LIST_CURSOR';
+import {
+    LOAD_QUERIES_LIST_ERROR,
+    LOAD_QUERIES_LIST_REQUEST,
+    LOAD_QUERIES_LIST_SUCCESS,
+    QueriesHistoryCursorDirection,
+    SET_QUERIES_LIST_CURSOR,
+    SET_QUERIES_LIST_FILTER,
+    SET_QUERIES_LIST_MODE,
+    UPDATE_QUERIES_LIST,
+} from '../query-tracker-contants';
 
 export type QueriesListAction =
     | Action<typeof LOAD_QUERIES_LIST_REQUEST>
