+++ packages/ui/src/ui/pages/query-tracker/module/query_aco/actions.ts
@@ -1,24 +1,13 @@
-import {Action} from 'redux';
 import {ThunkAction} from 'redux-thunk';
 import {Toaster} from '@gravity-ui/uikit';
 import type {AxiosError} from 'axios';
 import {YTApiId, ytApiV4Id} from '../../../../rum/rum-wrap-api';
 import {getQueryTrackerRequestOptions} from '../query/selectors';
 import {QUERY_ACO_LOADING} from './constants';
-import type {ActionD} from '../../../../types';
 import {showErrorPopup} from '../../../../utils/utils';
-import type {QueryACOState} from './reducer';
+import {QueryACOActions} from './reducer';
 
-export type QueryACOLoadingAction = Action<typeof QUERY_ACO_LOADING.REQUEST>;
-export type QueryACOLoadingError = Action<typeof QUERY_ACO_LOADING.FAILURE>;
-export type QueryACOLoadingSuccess = ActionD<
-    typeof QUERY_ACO_LOADING.SUCCESS,
-    QueryACOState['data']
->;
-
-export type QueryACOActions = QueryACOLoadingAction | QueryACOLoadingError | QueryACOLoadingSuccess;
-
-export const getQueryACO = (): ThunkAction<Promise<unknown>, any, any, any> => {
+export const getQueryACO = (): ThunkAction<Promise<unknown>, any, any, QueryACOActions> => {
     return (dispatch, getState) => {
         const state = getState();
         const {stage} = getQueryTrackerRequestOptions(state);
